fleshed out bidName(string), refactored name handling, implemented cluster presence controls
bidName() was expanded to process bids if name is still owned after poke. if bid is existing, add to it. if name ownership was purged after poke, it now iterates through all bids and transfers the name to the highest bidder, regardless if they are msg.sender or a previous bidder. msg.sender's bid will be handled appropriately whether they are the highest bidder when placing their bid or not. _unassignName() was implemented so _transferName() could be refactored. now nameLookup is updated appropriately. lastly, buyName() and bidName() were restricted to senders who already created their own cluster. names cannot be bought or bid on if no cluster exists.