refactored pokeName() to transfer name to highest bidder, if any
bidName(string) used to call pokeName(string) at the beginning of its function and then process transferring the name to the highest bidder if pokeName() exhausted all remaining ETH backing. now pokeName() processes the transfer to highest bidder and bidName() stores the bid made in the caller's context before calling pokeName(). The redesigned _transferName() function allows for poke to use it to delete or transfer the name, regardless of bid status.