collapsed invite+join flow to add-only
soft additions are optimal as it allows one wallet to add multiple addresses in a single multicall and process the change instead of having to accept from each wallet. economic incentives keep this in check as any address can kick out any other address, so trying to grief or impersonate addresses by adding them to a cluster is to the detriment of the cluster owner. better UX is better than the tradeoff here due to equal permissions across all cluster addresses.