Fix interface (#13)
Large amounts of the test suite are failing because it attempts to bid a name that doesn't exist. After removing the clusterId parameter from bidName (which is desired behavior), need to rewrite these tests to not bid on names that don't exist yet. Should only buy them directly if nobody owns them.