added all necessary function overloads for payable multicall
in order to ensure our invariants hold true, I've made overloads for all non-payable functions so they can be called in a payable multicall. non-payable functions cannot be called when there is msg.value, so if any of the calls in the multicall is to a payable function, these overloads enable handling non-payment to originally unpayable functions. all multicall overloads are prefaced with a value parameter. buyName, fundName, and bidName are the only functions that use it as they are the only truly payable functions. all other functions discard this parameter entirely and simply use it to help identify multicallable payable overrides.