realized ClustersHubRelay.sol was unnecessary (...)
ClustersHubRelay.sol is basically ClustersSpoke.sol. There was no reason for a message to pass to the Relay's endpoint, to the Relay itself, and then back to the endpoint. The endpoint itself is what is responsible for state duplication, so the functionality for it should be built there. Also renamed NameManagerMain.sol to NameManagerHub.sol to fit the new naming convention.